drv: update header and kline impl

+++ drivers/j2534/j2534/j2534translate.h
@@ -8,11 +8,11 @@
 #ifndef J2534TRANSLATE_H_
 #define J2534TRANSLATE_H_
 
-char* translateError(uint8_t err);
+char* translateError(uint32_t err);
 char* translateErrorDetail(uint32_t err);
-char* translateIoctl(uint8_t ioctl);
-char* translateParam(uint8_t param);
-char* translateProtocol(uint8_t protocol);
+char* translateIoctl(uint32_t ioctl);
+char* translateParam(uint32_t param);
+char* translateProtocol(uint32_t protocol);
 char* translateFilterType(uint32_t type);
 
 #endif /* J2534TRANSLATE_H_ */
